fix sometimes poem is missing
I added some code that hides the poem when it's connected.
but this variable isn't properly set when you, for example,
press back and forward. the poem is just missing.

so I removed the code that hides the poem. this makes the page a little
uglier when it's loading (for some reason all the word selectors are
open) but this prevents the page from breaking more than it needs to