fix: prevent word lists from being open as the page loads
the logic to determine if a word list should be open is simply checking
if a given blank is equal to the user's id.

these menus were popping open because when both of these values are
undefined, they equal each other.

i fixed this by checking to make sure the id isn't undefined