fix: FillingIn bug in production mode
the code i used to look up the index of a word selector was breaking
when running in production mode. Instead of debugging the issue, I
hardcoded a prop for each word selector so it knows which one it is. a
bit of a hacky solution, but it seems to fix the issue, allowing us to
run in production mode.

This also changes how we loop through all wordselectors to add words
whenever blanklist is updated.